cloudvisionai-backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.4.0
gunicorn==22.0.0